Sostituisco gli INT con i FLOAT

+++ src/main/java/java_calculator_tester/Calc.java
@@ -2,19 +2,19 @@ package java_calculator_tester;
 
 public class Calc {
 
-	public int add(int num1, int num2) {
+	public float add(float num1, float num2) {
         return num1 + num2;
     }
 
-    public int subtract(int num1, int num2) {
+    public float subtract(float num1, float num2) {
         return num1 - num2;
     }
 
-    public int divide(int num1, int num2) {
+    public float divide(float num1, float num2) {
         return num1 / num2;
     }
 
-    public int multiply(int num1, int num2) {
+    public float multiply(float num1, float num2) {
         return num1 * num2;
     }
 }
